scripts/pymledger/requirements.txt
setuptools>=60.9
pyyaml>=6.0
colorama>=0.4
docopt>=0.6
pandas>=1.4
plotly>=5.6
kaleido>=0.2